Double check that the file exists before attempting to remove.

+++ library/src/main/java/com/novoda/downloadmanager/ExternalFilePersistence.java
@@ -134,6 +134,10 @@ class ExternalFilePersistence implements FilePersistence {
         }
 
         File fileToDelete = new File(absoluteFilePath.path());
+        if (!fileToDelete.exists()) {
+            return;
+        }
+
         boolean deleted = fileToDelete.delete();
 
         String message = String.format("File or Directory: %s deleted: %s", absoluteFilePath.path(), deleted);
